Config reading now Works.
