--- src/main/java/ch/hutch79/configManager/Configuration.java
@@ -1,13 +1,0 @@
-package ch.hutch79.configManager;
-
-public class Configuration {
-    private int version;
-
-    public int getVersion() {
-        return version;
-    }
-
-    public void setVersion(int version) {
-        this.version = version;
-    }
-}
