+++ src/main/java/ch/hutch79/configManager/configClass/config/v0/Config.java
@@ -1,0 +1,27 @@
+package ch.hutch79.configManager.configClass.config.v0;
+
+
+import java.util.List;
+
+public class Config {
+    private boolean debug;
+    private List<Command> command;
+
+
+
+    public boolean getDebug() {
+        return debug;
+    }
+
+    public void setDebug(boolean debug) {
+        this.debug = debug;
+    }
+
+    public List<Command> getCommand() {
+        return command;
+    }
+
+    public void setCommand(List<Command> command) {
+        this.command = command;
+    }
+}
